refactor(*.*): 패키지명, 프로젝트 네임 수정
