+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'jordy-shop'
+rootProject.name = 'delivery-saurus'
