+++ src/main/java/com/deliverysaurus/DeliverySaurusApplication.java
@@ -1,13 +1,13 @@
-package com.example.jordyshop;
+package com.deliverysaurus;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class JordyShopApplication {
+public class DeliverySaurusApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(JordyShopApplication.class, args);
+		SpringApplication.run(DeliverySaurusApplication.class, args);
 	}
 
 }
