+++ src/main/java/com/deliverysaurus/domain/Status.java
@@ -1,4 +1,4 @@
-package com.example.jordyshop.domain;
+package com.deliverysaurus.domain;
 
 import javax.persistence.Embeddable;
