+++ src/test/java/com/deliverysaurus/domain/MemberTest.java
@@ -1,4 +1,4 @@
-package com.example.jordyshop.domain;
+package com.deliverysaurus.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
